defect fix: deviceWidth needs to be handled as a double!

+++ app/models/UserSettings.scala
@@ -20,10 +20,10 @@ object Cuisine {
 case class UserSettings(id: Long, var language: String, skipDailyEmail: Boolean, 
     timeZone: String, screenWidth: Long, var favCuisines: MutableList[Cuisine], 
     var preferToAvoid: Option[MutableList[Cuisine]], var sampleDishLikes: Option[MutableList[Cuisine]],
-    var deviceOS: Option[String], var deviceSWidth: Option[Long], var deviceLang: Option[String])
+    var deviceOS: Option[String], var deviceSWidth: Option[Double], var deviceLang: Option[String])
     
 object UserSettings {
-  def default(id: Long) = { new UserSettings(id, "en_US", false, "", 0, MutableList.empty[Cuisine], Option(MutableList.empty[Cuisine]), Option(MutableList.empty[Cuisine]), Option(""), Option(0), Option("")) }
+  def default(id: Long) = { new UserSettings(id, "en_US", false, "", 0, MutableList.empty[Cuisine], Option(MutableList.empty[Cuisine]), Option(MutableList.empty[Cuisine]), Option(""), Option(0.0), Option("")) }
   
   implicit val userSettingsReads = Json.reads[UserSettings]
   implicit val userSettingsWrites = Json.writes[UserSettings]
